flow: split python-formatting dependant nodes into their own module
Let's call it format and have aliases back to the original
namespaces, namely form and converter.
