+++ src/lib/flow/Kconfig
@@ -59,6 +59,7 @@ source "src/modules/flow/filter-repeated/Kconfig"
 source "src/modules/flow/float/Kconfig"
 source "src/modules/flow/flower-power/Kconfig"
 source "src/modules/flow/form/Kconfig"
+source "src/modules/flow/format/Kconfig"
 source "src/modules/flow/gpio/Kconfig"
 source "src/modules/flow/grove/Kconfig"
 source "src/modules/flow/gtk/Kconfig"
@@ -67,9 +68,9 @@ source "src/modules/flow/http-client/Kconfig"
 source "src/modules/flow/http-server/Kconfig"
 source "src/modules/flow/iio/Kconfig"
 source "src/modules/flow/int/Kconfig"
+source "src/modules/flow/jhd1313m1/Kconfig"
 source "src/modules/flow/json/Kconfig"
 source "src/modules/flow/keyboard/Kconfig"
-source "src/modules/flow/jhd1313m1/Kconfig"
 source "src/modules/flow/led-7seg/Kconfig"
 source "src/modules/flow/led-strip/Kconfig"
 source "src/modules/flow/location/Kconfig"
