+++ src/modules/flow/converter/Makefile
@@ -1,3 +1,3 @@
 obj-$(FLOW_NODE_TYPE_CONVERTER) += converter.mod
-obj-converter-$(FLOW_NODE_TYPE_CONVERTER) := converter.json converter.o string-format.o
+obj-converter-$(FLOW_NODE_TYPE_CONVERTER) := converter.json converter.o
 obj-converter-$(FLOW_NODE_TYPE_CONVERTER)-type := flow
