+++ src/modules/flow/form/Makefile
@@ -1,6 +1,3 @@
 obj-$(FLOW_NODE_TYPE_FORM) += form.mod
-obj-form-$(FLOW_NODE_TYPE_FORM) := form.json form.o ../converter/string-format.o ../converter/string-format.h
+obj-form-$(FLOW_NODE_TYPE_FORM) := form.json form.o form-common.o
 obj-form-$(FLOW_NODE_TYPE_FORM)-type := flow
-obj-form-$(FLOW_NODE_TYPE_FORM)-deps := flow/converter.mod
-obj-form-$(FLOW_NODE_TYPE_FORM)-extra-ldflags := \
-        -lm
