+++ src/modules/flow/form/form.c
@@ -20,21 +20,12 @@
 #include <stdio.h>
 
 #include "sol-flow/form.h"
-#include "../converter/string-format.h"
 #include "sol-buffer.h"
 #include "sol-mainloop.h"
+#include "sol-flow-internal.h"
 #include "sol-util-internal.h"
 
-#define DITCH_NL (true)
-#define KEEP_NL (false)
-
-#define DO_FORMAT (false)
-#define CALC_ONLY (true)
-
-static const char CR = '\r', NL = '\n', SPC = ' ', UNDERSCORE = '_',
-    NUL = '\0', CURL_BRACKET_OPEN = '{', CURL_BRACKET_CLOSE = '}', COMMA = ',';
-static const char TITLE_TAG[] = "{title}", VALUE_TAG[] = "{value}",
-    EMPTY_STR[] = "";
+#include "form-common.h"
 
 struct selector_data {
     char *title, *sel_mark, *cursor_mark, *text_mem;
@@ -47,295 +38,6 @@ struct selector_data {
     bool enabled : 1;
     bool n_values_done : 1;
 };
-
-static inline int
-get_buf_size(size_t rows, size_t columns, size_t *out)
-{
-    size_t n_cols;
-    int r;
-
-    /* +1 on cols for '\n' and (final) NUL chars */
-    r = sol_util_size_add(columns, 1, &n_cols);
-    SOL_INT_CHECK(r, < 0, r);
-
-    r = sol_util_size_mul(rows, n_cols, out);
-    SOL_INT_CHECK(r, < 0, r);
-
-    return 0;
-}
-
-static inline int
-buffer_re_init(struct sol_buffer *buf,
-    char *mem,
-    size_t rows,
-    size_t columns)
-{
-    size_t size;
-
-    int r = get_buf_size(rows, columns, &size);
-
-    SOL_INT_CHECK(r, < 0, r);
-
-    /* We choose to do the ending nul byte ourselves, since we're
-     * doing fixed capacity, to be able to set the second to last byte
-     * on the buffer without reallocation attempts. */
-    sol_buffer_init_flags(buf, mem, size,
-        SOL_BUFFER_FLAGS_FIXED_CAPACITY | SOL_BUFFER_FLAGS_NO_NUL_BYTE);
-
-    return 0;
-}
-
-static int
-common_form_init(struct sol_buffer *buf,
-    int32_t in_rows,
-    size_t *out_rows,
-    int32_t in_cols,
-    size_t *out_cols,
-    const char *in_format,
-    char **out_format,
-    const char *in_title,
-    char **out_title,
-    char **out_title_tag,
-    char **out_value_tag,
-    char **out_text_mem)
-{
-    int r = 0;
-    size_t size;
-
-    SOL_NULL_CHECK(in_format, -EINVAL);
-
-    if (in_rows <= 0) {
-        SOL_WRN("Form rows number must be a positive integer, "
-            "but %" PRId32 " was given. Fallbacking to minimum value of 1.",
-            in_rows);
-        *out_rows = 1;
-    } else
-        *out_rows = in_rows;
-
-    if (in_cols <= 0) {
-        SOL_WRN("Boolean columns number must be a positive integer, "
-            "but %" PRId32 " was given. Fallbacking to minimum value of 1.",
-            in_cols);
-        *out_cols = 1;
-    } else
-        *out_cols = in_cols;
-
-    r = get_buf_size(*out_rows, *out_cols, &size);
-    SOL_INT_CHECK(r, < 0, r);
-
-    *out_text_mem = calloc(1, size);
-    SOL_NULL_CHECK(*out_text_mem, -ENOMEM);
-
-    *out_format = strdup(in_format);
-    if (!*out_format) {
-        r = -ENOMEM;
-        goto err_format;
-    }
-
-    *out_title_tag = strstr(*out_format, TITLE_TAG);
-    *out_value_tag = strstr(*out_format, VALUE_TAG);
-
-    if (!*out_value_tag) {
-        SOL_WRN("Bad format, no {value} tag: %s. Fallbacking to "
-            "pristine one, i. e. '{value}'.", *out_format);
-        r = sol_util_replace_str_if_changed(out_format, VALUE_TAG);
-        *out_value_tag = *out_format;
-        *out_title_tag = NULL;
-        SOL_INT_CHECK_GOTO(r, < 0, err_tags);
-    }
-
-    if (*out_title_tag > *out_value_tag) {
-        SOL_WRN("Bad format, {title} tag placed after {value} tag: %s."
-            " Fallbacking to pristine one, i. e. '{value}'.",
-            *out_format);
-        r = sol_util_replace_str_if_changed(out_format, VALUE_TAG);
-        *out_value_tag = *out_format;
-        *out_title_tag = NULL;
-        SOL_INT_CHECK_GOTO(r, < 0, err_tags);
-    }
-
-    if (in_title && *out_title_tag) {
-        *out_title = strdup(in_title);
-        if (!*out_title) {
-            r = -ENOMEM;
-            goto err_title;
-        }
-    } else {
-        *out_title = NULL;
-    }
-
-    r = buffer_re_init(buf, *out_text_mem, *out_rows, *out_cols);
-    if (r < 0) {
-        sol_buffer_fini(buf);
-        goto err_tags;
-    }
-
-    return 0;
-
-err_tags:
-err_title:
-    free(*out_format);
-
-err_format:
-    free(*out_text_mem);
-
-    return r;
-}
-
-static inline size_t
-coords_to_pos(size_t n_cols, size_t r, size_t c)
-{
-    /* account for extra col for (implicit) trailing NLs, thus + r */
-    return (r * n_cols) + c + r;
-}
-
-#define CUR_POS (coords_to_pos(n_cols, *row, *col))
-#define CUR_EXTRA_COL (coords_to_pos(n_cols, *row, n_cols - 1) + 1)
-
-static int
-fill_spaces(struct sol_buffer *buf,
-    size_t n_cols,
-    size_t *row,
-    size_t *col,
-    size_t length)
-{
-    int r;
-
-    while (*col < n_cols && length) {
-        r = sol_buffer_set_char_at(buf, CUR_POS, SPC);
-        SOL_INT_CHECK(r, < 0, r);
-        (*col)++;
-        length--;
-    }
-    /* If we did not reach the row's end, we're done, otherwise we
-     * must break the line */
-    if (!length)
-        return CUR_POS;
-    r = sol_buffer_set_char_at(buf, CUR_EXTRA_COL, NL);
-    SOL_INT_CHECK(r, < 0, r);
-    *col = 0;
-    (*row)++;
-
-    return CUR_POS;
-}
-
-static int
-fill_line(struct sol_buffer *buf,
-    size_t n_rows,
-    size_t n_cols,
-    size_t *row,
-    size_t *col,
-    bool calc_only)
-{
-    int r;
-
-    while (*col < n_cols) {
-        if (!calc_only) {
-            r = sol_buffer_set_char_at(buf, CUR_POS, SPC);
-            SOL_INT_CHECK(r, < 0, r);
-        }
-        (*col)++;
-    }
-    if (*row < n_rows - 1) {
-        if (!calc_only) {
-            r = sol_buffer_set_char_at
-                    (buf, CUR_EXTRA_COL, NL);
-            SOL_INT_CHECK(r, < 0, r);
-        }
-    }
-    (*row)++;
-    *col = 0;
-
-    return CUR_POS;
-}
-
-static int
-go_to_new_line(struct sol_buffer *buf,
-    size_t n_rows,
-    size_t n_cols,
-    size_t *row,
-    size_t *col)
-{
-    int r = 0;
-
-    if (*col > 0)
-        r = fill_line(buf, n_rows, n_cols, row, col, DO_FORMAT);
-
-    return r;
-}
-
-static int
-format_chunk(struct sol_buffer *buf,
-    size_t n_rows,
-    size_t n_cols,
-    const char *ptr,
-    const char *end_ptr,
-    size_t *row,
-    size_t *col,
-    bool calc_only,
-    bool ditch_new_lines)
-{
-    size_t sz = buf->capacity;
-    int r;
-
-    while (ptr < end_ptr && CUR_POS < sz && *row < n_rows) {
-        /* translate middle-of-line \n's to spaces till the end +
-         * \n */
-        if (*ptr == CR || *ptr == NL) {
-            if (ditch_new_lines) {
-                if (!calc_only) {
-                    r = sol_buffer_set_char_at(buf, CUR_POS, SPC);
-                    SOL_INT_CHECK(r, < 0, r);
-                }
-                (*col)++;
-            } else {
-                r = fill_line(buf, n_rows, n_cols, row, col, calc_only);
-                SOL_INT_CHECK(r, < 0, r);
-            }
-            if (*ptr == CR)
-                ptr++;
-            if (*ptr == NL)
-                ptr++;
-            goto col_check;
-        }
-        if (!calc_only) {
-            r = sol_buffer_set_char_at(buf, CUR_POS, *ptr);
-            SOL_INT_CHECK(r, < 0, r);
-        }
-        (*col)++;
-        ptr++;
-
-col_check:
-        /* crop lines that don't fit */
-        if (*col == n_cols) {
-            if (!calc_only) {
-                if (*row < n_rows - 1) {
-                    r = sol_buffer_set_char_at(buf, CUR_EXTRA_COL, NL);
-                    SOL_INT_CHECK(r, < 0, r);
-                }
-            }
-            *col = 0;
-            (*row)++;
-            while (*ptr && *ptr != CR && *ptr != NL) {
-                ptr++;
-            }
-            if (*ptr == CR)
-                ptr++;
-            if (*ptr == NL)
-                ptr++;
-            if (ditch_new_lines) /* if ditching NL's, stop at 1st crop */
-                break;
-        } else if (ptr == end_ptr && !ditch_new_lines) {
-            /* the source ended before the line */
-            r = fill_line(buf, n_rows, n_cols, row, col, calc_only);
-            SOL_INT_CHECK(r, < 0, r);
-        }
-    }
-
-    return CUR_POS;
-#undef CUR_POS
-#undef CUR_EXTRA_COL
-}
 
 static int
 calculate_n_values(struct selector_data *mdata, size_t *row, size_t *col)
@@ -364,110 +66,6 @@ calculate_n_values(struct selector_data *mdata, size_t *row, size_t *col)
     mdata->n_values_done = true;
 
     return 0;
-}
-
-static int
-format_title(struct sol_buffer *buf,
-    size_t buf_size,
-    size_t n_rows,
-    size_t n_cols,
-    size_t *row_ptr,
-    size_t *col_ptr,
-    const char *format,
-    const char *title,
-    const char *title_tag,
-    const char *value_tag,
-    bool *no_more_space)
-{
-    int r;
-    const char *ptr = format;
-
-    *no_more_space = false;
-
-    /* Format pre-title/value chunk */
-    r = format_chunk(buf, n_rows, n_cols, ptr,
-        title_tag ? title_tag : value_tag,
-        row_ptr, col_ptr, DO_FORMAT, KEEP_NL);
-    SOL_INT_CHECK(r, < 0, r);
-    if (r >= (int)buf_size || *row_ptr >= n_rows) {
-        *no_more_space = true;
-        return 0;
-    }
-
-    if (!title || !title_tag)
-        return 0;
-
-    /* Format title */
-    r = format_chunk(buf, n_rows, n_cols, title, title + strlen(title),
-        row_ptr, col_ptr, DO_FORMAT, DITCH_NL);
-    SOL_INT_CHECK(r, < 0, r);
-    if (r >= (int)buf_size || *row_ptr >= n_rows) {
-        *no_more_space = true;
-        return 0;
-    }
-
-    if (n_rows > 1) {
-        r = go_to_new_line(buf, n_rows, n_cols, row_ptr, col_ptr);
-        SOL_INT_CHECK(r, < 0, r);
-    } else {
-        r = fill_spaces(buf, n_cols, row_ptr, col_ptr, 1);
-        SOL_INT_CHECK(r, < 0, r);
-    }
-
-    /* Format post-title, pre-value chunk. If we got only one
-     * line, ditch new lines in an attempt to have both title and
-     * value in it */
-    ptr = title_tag + strlen(TITLE_TAG);
-    r = format_chunk(buf, n_rows, n_cols, ptr, value_tag, row_ptr,
-        col_ptr, DO_FORMAT, n_rows > 1 ? KEEP_NL : DITCH_NL);
-    SOL_INT_CHECK(r, < 0, r);
-    if (r >= (int)buf_size || *row_ptr >= n_rows)
-        *no_more_space = true;
-
-    return 0;
-}
-
-static int
-format_post_value(struct sol_buffer *buf,
-    size_t n_rows,
-    size_t n_cols,
-    size_t *row_ptr,
-    size_t *col_ptr,
-    const char *format,
-    const char *value_tag)
-{
-    const char *ptr = value_tag + sizeof(VALUE_TAG) - 1;
-    int r;
-
-    r = go_to_new_line(buf, n_rows, n_cols, row_ptr, col_ptr);
-    SOL_INT_CHECK(r, < 0, r);
-
-    /* Format post-value chunk */
-    return format_chunk(buf, n_rows, n_cols, ptr,
-        format + strlen(format), row_ptr, col_ptr, DO_FORMAT, KEEP_NL);
-}
-
-static int
-format_send(struct sol_flow_node *node,
-    struct sol_buffer *buf,
-    uint16_t out_port)
-{
-    int r;
-
-    /* Don't ever end with NL and guarantee ending NUL byte */
-    if (buf->used) {
-        char *value = (char *)sol_buffer_at_end(buf);
-        *value = NUL;
-        value = (char *)sol_buffer_at_end(buf) - 1;
-        if (*value == NL)
-            *value = NUL;
-    } else {
-        r = sol_buffer_set_char_at(buf, 0, NUL);
-        SOL_INT_CHECK(r, < 0, r);
-    }
-
-    return sol_flow_send_string_slice_packet(node,
-        out_port, sol_buffer_get_slice(buf));
 }
 
 //FIXME: - autoscroll/markee effect on tags
@@ -2555,968 +2153,6 @@ send:
 
     string_force_imediate_format(mdata, true);
     return string_format(node);
-}
-
-struct string_formatted_data {
-    size_t columns, rows, value_prefix_len, cursor;
-    char *title, *text_mem, *format, *value, *title_tag, *value_tag;
-    struct sol_buffer text_grid, formatted_value;
-    struct sol_vector chunks;
-    struct sol_timeout *timer;
-    int blink_time;
-    bool circular : 1;
-    bool enabled : 1;
-    bool blink_on : 1;
-    bool state_changed : 1;
-    bool cursor_initialized : 1;
-};
-
-enum string_formatted_chunk_type {
-    STR_FORMAT_INT,
-    STR_FORMAT_FLOAT,
-    STR_FORMAT_LITERAL
-};
-
-struct string_formatted_chunk {
-    /* Slice from mdata->formatted_value.data, on INT/FLOAT types, or
-     * from mdata->value, for LITERALs. We have to keep this to have
-     * the information again on the blinking process (for INT/FLOAT)
-     * or to easily resolve the LITERAL when formatting. */
-    struct sol_str_slice rendered;
-
-    /* Re-built from mdata->value, on INT/FLOAT types, to match
-     * do_{integer,float}_markup()'s syntax */
-    char *format;
-
-    /* Starting position at mdata->text_grid.data, on INT/FLOAT types.
-     * If the field does not fit at all, it will contain NULL */
-    const char *pos_in_text_grid;
-
-    enum string_formatted_chunk_type type;
-
-    /* Numerical state, for INT/FLOAT types */
-    union {
-        struct sol_irange i;
-        struct sol_drange d;
-    } state;
-};
-
-static int
-string_formatted_format_do(struct sol_flow_node *node)
-{
-    struct string_formatted_data *mdata = sol_flow_node_get_private_data(node);
-    int r, buf_size = mdata->text_grid.capacity;
-    size_t row = 0, col = 0, tmp_col;
-    struct sol_vector indexes;
-    bool no_more_space = false;
-    char *tmp = NULL;
-
-    if (!mdata->state_changed) {
-        char *value;
-        size_t start, len;
-        struct string_formatted_chunk *chunk =
-            sol_vector_get(&mdata->chunks, mdata->cursor);
-
-        if (!chunk->pos_in_text_grid)
-            goto send;
-
-        start = chunk->pos_in_text_grid - (char *)mdata->text_grid.data;
-
-        /* abs. pos modulo number of cols gives us the actual col */
-        len = mdata->columns - (start % mdata->columns) + 1;
-
-        value = (char *)sol_buffer_at(&mdata->text_grid, start);
-
-        if (mdata->blink_on) {
-            mdata->blink_on = false;
-            for (size_t l = 0; l < chunk->rendered.len && len; l++, len--)
-                value[l] = SPC;
-        } else {
-            mdata->blink_on = true;
-            for (size_t l = 0; l < chunk->rendered.len && len; l++, len--)
-                value[l] = chunk->rendered.data[l];
-        }
-
-        goto send;
-    }
-
-    sol_vector_init(&indexes, sizeof(size_t));
-
-    r = format_title(&mdata->text_grid, buf_size, mdata->rows,
-        mdata->columns, &row, &col, mdata->format, mdata->title,
-        mdata->title_tag, mdata->value_tag, &no_more_space);
-    SOL_INT_CHECK_GOTO(r, < 0, err);
-    SOL_EXP_CHECK_GOTO(no_more_space, send);
-
-    mdata->value_prefix_len = tmp_col = col;
-
-    sol_buffer_fini(&mdata->formatted_value);
-    sol_buffer_init(&mdata->formatted_value);
-
-    for (size_t l = 0; l < mdata->chunks.len; l++) {
-        struct string_formatted_chunk *chunk =
-            sol_vector_get(&mdata->chunks, l);
-        size_t start, sz, *start_ptr;
-
-        start = mdata->formatted_value.used;
-
-        switch (chunk->type) {
-        case STR_FORMAT_INT:
-            r = do_integer_markup(node, chunk->format, chunk->state.i,
-                &mdata->formatted_value);
-            break;
-        case STR_FORMAT_FLOAT:
-            r = do_float_markup(node, chunk->format, chunk->state.d,
-                &mdata->formatted_value);
-            break;
-        case STR_FORMAT_LITERAL:
-            r = sol_buffer_append_slice
-                    (&mdata->formatted_value, chunk->rendered);
-            break;
-        }
-        SOL_INT_CHECK_GOTO(r, < 0, err);
-        sz = mdata->formatted_value.used - start;
-
-        if (chunk->type == STR_FORMAT_LITERAL)
-            continue;
-
-        /* We can't assume mdata->formatted_value.data + start will
-         * stay as is, because the buffer may grow on these iterations
-         * and be realloc()ed! We save the value to compute the ptr
-         * offsets later */
-        start_ptr = sol_vector_append(&indexes);
-        if (!start_ptr) {
-            r = -ENOMEM;
-            goto err;
-        }
-
-        *start_ptr = start;
-        chunk->rendered.len = sz;
-
-        if (tmp_col /* + sz */ > mdata->columns - 1)
-            chunk->pos_in_text_grid = NULL;
-        else {
-            chunk->pos_in_text_grid = (char *)mdata->text_grid.data +
-                coords_to_pos(mdata->columns, row, mdata->value_prefix_len)
-                + start;
-            tmp_col += sz;
-        }
-
-        if (!mdata->cursor_initialized) {
-            mdata->cursor = l;
-            mdata->cursor_initialized = true;
-        }
-    }
-
-    for (size_t l = 0, s = 0; l < mdata->chunks.len; l++) {
-        size_t *start_ptr;
-        struct string_formatted_chunk *chunk =
-            sol_vector_get(&mdata->chunks, l);
-
-        if (chunk->type == STR_FORMAT_LITERAL)
-            continue;
-
-        start_ptr = sol_vector_get(&indexes, s++);
-        if (!start_ptr) {
-            r = -ERANGE;
-            goto err;
-        }
-
-        chunk->rendered.data =
-            (char *)mdata->formatted_value.data + *start_ptr;
-    }
-
-    sol_vector_clear(&indexes);
-
-    tmp = mdata->formatted_value.data;
-    r = format_chunk(&mdata->text_grid, mdata->rows,
-        mdata->columns, tmp, tmp + mdata->formatted_value.used, &row, &col,
-        DO_FORMAT, DITCH_NL);
-    SOL_INT_CHECK_GOTO(r, < 0, err);
-    if (r >= buf_size || row >= mdata->rows)
-        goto send;
-
-    r = format_post_value(&mdata->text_grid, mdata->rows,
-        mdata->columns, &row, &col, mdata->format,
-        mdata->value_tag);
-    SOL_INT_CHECK_GOTO(r, < 0, err);
-
-    mdata->state_changed = false;
-
-send:
-    r = format_send(node, &mdata->text_grid,
-        SOL_FLOW_NODE_TYPE_FORM_STRING_FORMATTED__OUT__STRING);
-
-    return r;
-
-err:
-    /* we got to re-init because of the error cases. if this function
-     * fails we're no better, so don't check. */
-    buffer_re_init(&mdata->text_grid, mdata->text_mem,
-        mdata->rows, mdata->columns);
-
-    sol_buffer_fini(&mdata->formatted_value);
-    sol_vector_clear(&indexes);
-
-    return r;
-}
-
-static bool
-string_formatted_timeout(void *data)
-{
-    struct string_formatted_data *mdata = sol_flow_node_get_private_data(data);
-
-    if (!string_formatted_format_do(data)) {
-        mdata->timer = NULL;
-        return true;
-    }
-
-    return false;
-}
-
-static void
-string_formatted_force_imediate_format(struct string_formatted_data *mdata,
-    bool re_init)
-{
-    if (re_init)
-        buffer_re_init(&mdata->text_grid, mdata->text_mem,
-            mdata->rows, mdata->columns);
-    if (mdata->timer) {
-        sol_timeout_del(mdata->timer);
-        mdata->timer = NULL;
-    }
-}
-
-static int
-string_formatted_format(struct sol_flow_node *node)
-{
-    struct string_formatted_data *mdata = sol_flow_node_get_private_data(node);
-
-    if (!mdata->timer) {
-        mdata->timer = sol_timeout_add
-                (mdata->blink_time, string_formatted_timeout, node);
-        SOL_NULL_CHECK(mdata->timer, -ENOMEM);
-        return string_formatted_format_do(node);
-    }
-
-    return 0;
-}
-
-static void
-string_formatted_close(struct sol_flow_node *node, void *data)
-{
-    struct string_formatted_data *mdata = data;
-    struct string_formatted_chunk *chunk;
-    uint16_t i;
-
-    sol_buffer_fini(&mdata->text_grid);
-    sol_buffer_fini(&mdata->formatted_value);
-
-    SOL_VECTOR_FOREACH_IDX (&mdata->chunks, chunk, i)
-        free(chunk->format);
-
-    sol_vector_clear(&mdata->chunks);
-
-    if (mdata->timer)
-        sol_timeout_del(mdata->timer);
-
-    free(mdata->title);
-    free(mdata->format);
-    free(mdata->value);
-}
-
-static inline double
-strtod_no_locale(const char *nptr, char **endptr)
-{
-    return sol_util_strtodn(nptr, endptr, -1, false);
-}
-
-static double
-midpoint(double min, double max)
-{
-    if (min < 0 && max > 0)
-        return (max + min) / 2.0;
-
-    return ((max - min) / 2.0) + min;
-}
-
-static int
-string_formatted_selected_set_do(struct sol_flow_node *node,
-    void *data,
-    const char *value)
-{
-    struct string_formatted_data *mdata = data;
-    const char *ptr, *end;
-    size_t len;
-    int r;
-
-    len = strlen(value);
-    ptr = value;
-    end = value + len;
-    for (size_t l = 0; l < mdata->chunks.len && *ptr; l++) {
-        struct string_formatted_chunk *chunk
-            = sol_vector_get(&mdata->chunks, l), *next_chunk;
-
-        switch (chunk->type) {
-        case STR_FORMAT_INT:
-        {
-            char *endptr;
-            int32_t *i_ptr;
-
-            if (l + 1 < mdata->chunks.len) {
-                next_chunk = sol_vector_get(&mdata->chunks, l + 1);
-
-                if (next_chunk->type != STR_FORMAT_LITERAL) {
-                    r = -ENOTSUP;
-                    goto error;
-                }
-                endptr = memmem(ptr, end - ptr,
-                    next_chunk->rendered.data, next_chunk->rendered.len);
-                if (!endptr) {
-                    r = -EINVAL;
-                    goto error;
-                }
-            } else
-                endptr = (char *)end;
-
-            i_ptr = &chunk->state.i.val;
-            *i_ptr = strtoll(ptr, &endptr, 0);
-
-            if (ptr == endptr || errno != 0) {
-                r = -errno;
-                goto error;
-            }
-            ptr = endptr;
-            break;
-        }
-        case STR_FORMAT_FLOAT:
-        {
-            char *endptr;
-            double *d_ptr;
-
-            if (l + 1 < mdata->chunks.len) {
-                next_chunk = sol_vector_get(&mdata->chunks, l + 1);
-
-                if (next_chunk->type != STR_FORMAT_LITERAL) {
-                    r = -ENOTSUP;
-                    goto error;
-                }
-                endptr = memmem(ptr, end - ptr,
-                    next_chunk->rendered.data, next_chunk->rendered.len);
-                if (!endptr) {
-                    r = -EINVAL;
-                    goto error;
-                }
-            } else
-                endptr = (char *)end;
-
-            d_ptr = &chunk->state.d.val;
-            *d_ptr = strtod_no_locale(ptr, &endptr);
-
-            if (ptr == endptr || errno != 0) {
-                r = -errno;
-                goto error;
-            }
-            ptr = endptr;
-            break;
-        }
-        case STR_FORMAT_LITERAL:
-            if (chunk->rendered.len > (size_t)(end - ptr) ||
-                memcmp(chunk->rendered.data, ptr, chunk->rendered.len) != 0) {
-                r = -EINVAL;
-                goto error;
-            }
-            ptr += chunk->rendered.len;
-            break;
-        }
-    }
-
-    return 0;
-
-error:
-    sol_flow_send_error_packet(node, ENOTSUP, "The node's value formatting "
-        "string (%s) is so that this entry (%s) is not parseable: %s",
-        mdata->value, value, sol_util_strerrora(-r));
-    return r;
-}
-
-static int
-string_formatted_open(struct sol_flow_node *node,
-    void *data,
-    const struct sol_flow_node_options *options)
-{
-    int r;
-    const char *ptr;
-    bool numeric_field_present = false;
-    struct string_formatted_data *mdata = data;
-    static const char syntax_msg[] =
-        "Please use the {<type>:<min>,<max>,<step>} syntax.";
-    struct string_formatted_chunk *chunk = NULL;
-    const struct sol_flow_node_type_form_string_formatted_options *opts =
-        (const struct sol_flow_node_type_form_string_formatted_options *)options;
-
-    SOL_FLOW_NODE_OPTIONS_SUB_API_CHECK(options,
-        SOL_FLOW_NODE_TYPE_FORM_INT_OPTIONS_API_VERSION, -EINVAL);
-
-    if (!opts->value) {
-        SOL_WRN("A value format must be passed.");
-        return -EINVAL;
-    } else {
-        mdata->value = strdup(opts->value);
-        SOL_NULL_CHECK(mdata->value, -ENOMEM);
-    }
-    ptr = mdata->value;
-
-    r = common_form_init(&mdata->text_grid,
-        opts->rows,
-        &mdata->rows,
-        opts->columns,
-        &mdata->columns,
-        opts->format,
-        &mdata->format,
-        opts->title,
-        &mdata->title,
-        &mdata->title_tag,
-        &mdata->value_tag,
-        &mdata->text_mem);
-    SOL_INT_CHECK(r, < 0, r);
-
-    mdata->circular = opts->circular;
-
-    mdata->enabled = true;
-
-    mdata->blink_time = opts->blink_time;
-    if (opts->blink_time < 0) {
-        SOL_WRN("Invalid blink_time (%" PRId32 "), that must be positive. "
-            "Setting to 1ms.", opts->blink_time);
-        mdata->blink_time = 1;
-    }
-
-    mdata->blink_on = true;
-    mdata->state_changed = true;
-
-    sol_vector_init(&mdata->chunks, sizeof(struct string_formatted_chunk));
-
-    /* For instance, a format could be something like
-     * "LITERAL{3d:0,255,1}LITERAL{3d:0,255,1}LITERAL" */
-
-    if (*ptr == CURL_BRACKET_OPEN)
-        goto numeric_loop;
-
-    /* literals state loop */
-literal_loop:
-    chunk = sol_vector_append(&mdata->chunks);
-    if (!chunk) {
-        r = -errno;
-        goto value_err;
-    }
-
-    chunk->type = STR_FORMAT_LITERAL;
-    chunk->rendered.data = ptr;
-    chunk->pos_in_text_grid = NULL;
-
-    while (*ptr) {
-        if (*ptr != CURL_BRACKET_OPEN) {
-            chunk->rendered.len++;
-            ptr++;
-        } else
-            break;
-    }
-
-    /* numeric field state loop */
-numeric_loop:
-    while (*ptr) {
-        const char *field_start = ptr, *field_end, *field_format_end;
-        char *tmp;
-
-        if (*ptr != CURL_BRACKET_OPEN)
-            goto literal_loop;
-
-        tmp = strchr(ptr, ':');
-
-        if (!tmp) {
-            r = -EINVAL;
-            goto value_err;
-        }
-
-        /* at least one char between '{' and ':' */
-        if (tmp - ptr < 2) {
-            SOL_WRN("No numeric field type specification passed (%.*s). %s.",
-                (int)(tmp - ptr), ptr, syntax_msg);
-            r = -EINVAL;
-            goto value_err;
-        }
-
-        field_format_end = tmp - 1;
-
-        chunk = sol_vector_append(&mdata->chunks);
-        if (!chunk) {
-            r = -errno;
-            goto value_err;
-        }
-
-        switch (*(tmp - 1)) {
-        case 'b':
-        case 'c':
-        case 'd':
-        case 'o':
-        case 'x':
-        case 'X':
-        case 'n':
-            chunk->type = STR_FORMAT_INT;
-            break;
-        case 'e':
-        case 'E':
-        case 'f':
-        case 'F':
-        case 'g':
-        case 'G':
-        case '%':
-            chunk->type = STR_FORMAT_FLOAT;
-            break;
-        default:
-            SOL_WRN("Bad numeric type (%c) given on format string. It must be "
-                "one of b/c/d/o/x/X/n, for integers, or e/E/f/F/g/G/%%, for"
-                " floating point numbers.", *(tmp - 1));
-            r = -EINVAL;
-            goto value_err;
-        }
-
-        sol_buffer_init(&mdata->formatted_value);
-
-#define NUM_SPEC_FIELDS (3)
-        /* check for min, max, step */
-        for (size_t i = 0; i < NUM_SPEC_FIELDS; i++) {
-            char *endptr, backup;
-
-            ptr = tmp + 1;
-            if (!*ptr) {
-                r = -EINVAL;
-                SOL_WRN("Numeric field format ended with missing min, max, "
-                    "step triple (%.*s). %s.", (int)(ptr - field_start),
-                    field_start, syntax_msg);
-                goto value_err;
-            }
-
-            tmp = strchr(ptr,
-                i < NUM_SPEC_FIELDS - 1 ? COMMA : CURL_BRACKET_CLOSE);
-            if (!tmp) {
-                SOL_WRN("Numeric field format ended with missing min, max, "
-                    "step triple (%.s). %s.", field_start, syntax_msg);
-                r = -EINVAL;
-                goto value_err;
-            }
-            /* we edit the buffer to ease strtoll()'s job, but we can
-             * still free it */
-            backup = *tmp;
-            *tmp = NUL;
-#undef NUM_SPEC_FIELDS
-
-            errno = 0;
-            if (chunk->type == STR_FORMAT_INT) {
-                int32_t *i_ptr = &chunk->state.i.val + i + 1;
-                *i_ptr = strtoll(ptr, &endptr, 0);
-
-                if (ptr == endptr || errno != 0) {
-                    SOL_WRN("Failed to parse integer number %s: %s. %s.",
-                        ptr, errno ? sol_util_strerrora(errno) : "bad format",
-                        syntax_msg);
-                    r = -errno;
-                    *tmp = backup;
-                    goto value_err;
-                }
-
-                /* val = min, to start with */
-                if (!i) {
-                    i_ptr--;
-                    *i_ptr = i_ptr[1];
-                }
-            } else {
-                double *d_ptr = &chunk->state.d.val + i + 1;
-                *d_ptr = strtod_no_locale(ptr, &endptr);
-
-                if (ptr == endptr || errno != 0) {
-                    SOL_WRN("Failed to parse floating point number"
-                        " %s: %s. %s.", ptr,
-                        errno ? sol_util_strerrora(errno) : "bad format",
-                        syntax_msg);
-                    r = -errno;
-                    *tmp = backup;
-                    goto value_err;
-                }
-
-                /* val = min, to start with */
-                if (!i) {
-                    d_ptr--;
-                    *d_ptr = d_ptr[1];
-                }
-            }
-            *tmp = backup;
-        }
-        field_end = tmp;
-
-        if (chunk->type == STR_FORMAT_INT) {
-            int64_t range;
-
-            if (chunk->state.i.min > chunk->state.i.max) {
-                int32_t v;
-                SOL_WRN("Max value should be greater than "
-                    "min on %.*s. Swapping both values.",
-                    (int)(field_end - field_start + 1), field_start);
-                v = chunk->state.i.max;
-                chunk->state.i.max = chunk->state.i.min;
-                chunk->state.i.min = v;
-            }
-            range = (int64_t)chunk->state.i.max
-                - (int64_t)chunk->state.i.min;
-            if ((chunk->state.i.step > 0
-                && chunk->state.i.step > range)
-                || (chunk->state.i.step < 0
-                && chunk->state.i.step < -range)) {
-                SOL_WRN("Step value must fit the given range for %.*s."
-                    " Assuming 1 for it.",
-                    (int)(field_end - field_start + 1), field_start);
-                chunk->state.i.step = 1;
-            }
-        } else {
-            double mid_point, mid_step, mid_range;
-
-            if (chunk->state.d.min > chunk->state.d.max) {
-                double v;
-                SOL_WRN("Max value should be greater than "
-                    "min on %.*s. Swapping both values.",
-                    (int)(field_end - field_start + 1), field_start);
-                v = chunk->state.d.max;
-                chunk->state.d.max = chunk->state.d.min;
-                chunk->state.d.min = v;
-            }
-            mid_point = midpoint(chunk->state.d.min,
-                chunk->state.d.max);
-            mid_range = chunk->state.d.max - mid_point;
-            mid_step = chunk->state.d.step / 2.0;
-            if ((mid_step > 0 && mid_step > mid_range)
-                || (mid_step < 0 && mid_step < -mid_range)) {
-                SOL_WRN("Step value must fit the given range for %.*s."
-                    " Setting it to that exact range.",
-                    (int)(field_end - field_start + 1), field_start);
-                chunk->state.d.step = 2.0 * mid_step;
-            }
-        }
-
-        r = asprintf(&chunk->format, "{:%.*s}",
-            (int)(field_format_end - field_start), field_start + 1);
-        SOL_INT_CHECK_GOTO(r, < 0, value_err);
-
-        /* here we just validate de format string with initial 0
-         * state's value */
-        if (chunk->type == STR_FORMAT_INT)
-            r = do_integer_markup(NULL, chunk->format, chunk->state.i,
-                &mdata->formatted_value);
-        else
-            r = do_float_markup(NULL, chunk->format, chunk->state.d,
-                &mdata->formatted_value);
-
-        /* warning already issued, on errors */
-        SOL_INT_CHECK_GOTO(r, < 0, value_err);
-
-        numeric_field_present = true;
-
-        ptr = tmp + 1;
-        sol_buffer_fini(&mdata->formatted_value);
-    }
-
-    if (!numeric_field_present) {
-        SOL_WRN("At least one numeric field must occur in the value format"
-            " string (%s), but none was detected. %s at least once in that"
-            " format string.", mdata->value, syntax_msg);
-        r = -EINVAL;
-        goto value_err;
-    }
-
-    if (opts->value_default)
-        string_formatted_selected_set_do(node, data, opts->value_default);
-
-    return string_formatted_format(node);
-
-value_err:
-    string_formatted_close(node, mdata);
-    return r;
-}
-
-static int
-string_formatted_up_set(struct sol_flow_node *node,
-    void *data,
-    uint16_t port,
-    uint16_t conn_id,
-    const struct sol_flow_packet *packet)
-{
-    struct string_formatted_data *mdata = data;
-    struct string_formatted_chunk *chunk;
-
-    if (!mdata->enabled)
-        return 0;
-
-    chunk = sol_vector_get(&mdata->chunks, mdata->cursor);
-
-    if (chunk->type == STR_FORMAT_INT) {
-        if (chunk->state.i.step > 0) {
-            /* step > 0 && max - step > min, so no overflow */
-            if (chunk->state.i.val <= chunk->state.i.max
-                - chunk->state.i.step) {
-                chunk->state.i.val += chunk->state.i.step;
-            } else {
-                if (mdata->circular)
-                    chunk->state.i.val = chunk->state.i.min;
-            }
-        } else {
-            /* step < 0 && min - step > max, so no overflow */
-            if (chunk->state.i.val >= chunk->state.i.min
-                - chunk->state.i.step) {
-                chunk->state.i.val += chunk->state.i.step;
-            } else {
-                if (mdata->circular)
-                    chunk->state.i.val = chunk->state.i.max;
-            }
-        }
-    } else {
-        if (chunk->state.d.step > 0) {
-            /* step > 0 && max - step > min, so no overflow */
-            if (chunk->state.d.val <= chunk->state.d.max
-                - chunk->state.d.step) {
-                chunk->state.d.val += chunk->state.d.step;
-            } else {
-                if (mdata->circular)
-                    chunk->state.d.val = chunk->state.d.min;
-            }
-        } else {
-            /* step < 0 && min - step > max, so no overflow */
-            if (chunk->state.d.val >= chunk->state.d.min
-                - chunk->state.d.step) {
-                chunk->state.d.val += chunk->state.d.step;
-            } else {
-                if (mdata->circular)
-                    chunk->state.d.val = chunk->state.d.max;
-            }
-        }
-    }
-
-    mdata->state_changed = true;
-    mdata->blink_on = true;
-
-    string_formatted_force_imediate_format(mdata, true);
-    return string_formatted_format(node);
-}
-
-static int
-string_formatted_down_set(struct sol_flow_node *node,
-    void *data,
-    uint16_t port,
-    uint16_t conn_id,
-    const struct sol_flow_packet *packet)
-{
-    struct string_formatted_data *mdata = data;
-    struct string_formatted_chunk *chunk;
-
-    if (!mdata->enabled)
-        return 0;
-
-    chunk = sol_vector_get(&mdata->chunks, mdata->cursor);
-
-    if (chunk->type == STR_FORMAT_INT) {
-        if (chunk->state.i.step > 0) {
-            /* step > 0 && min + step < max, so no overflow */
-            if (chunk->state.i.val >= chunk->state.i.min
-                + chunk->state.i.step) {
-                chunk->state.i.val -= chunk->state.i.step;
-            } else {
-                if (mdata->circular)
-                    chunk->state.i.val = chunk->state.i.max;
-            }
-        } else {
-            /* step < 0 && max + step < min, so no overflow */
-            if (chunk->state.i.val <= chunk->state.i.max
-                + chunk->state.i.step) {
-                chunk->state.i.val -= chunk->state.i.step;
-            } else {
-                if (mdata->circular)
-                    chunk->state.i.val = chunk->state.i.min;
-            }
-        }
-    } else {
-        if (chunk->state.d.step > 0) {
-            /* step > 0 && min + step < max, so no overflow */
-            if (chunk->state.d.val >= chunk->state.d.min
-                + chunk->state.d.step) {
-                chunk->state.d.val -= chunk->state.d.step;
-            } else {
-                if (mdata->circular)
-                    chunk->state.d.val = chunk->state.d.max;
-            }
-        } else {
-            /* step < 0 && max + step < min, so no overflow */
-            if (chunk->state.d.val <= chunk->state.d.max
-                + chunk->state.d.step) {
-                chunk->state.d.val -= chunk->state.d.step;
-            } else {
-                if (mdata->circular)
-                    chunk->state.d.val = chunk->state.d.min;
-            }
-        }
-    }
-
-    mdata->state_changed = true;
-    mdata->blink_on = true;
-
-    string_formatted_force_imediate_format(mdata, true);
-    return string_formatted_format(node);
-}
-
-static int
-string_formatted_next_set(struct sol_flow_node *node,
-    void *data,
-    uint16_t port,
-    uint16_t conn_id,
-    const struct sol_flow_packet *packet)
-{
-    struct string_formatted_data *mdata = data;
-    size_t cursor_pos = mdata->cursor, l = mdata->chunks.len;
-    bool found = false;
-
-    if (!mdata->enabled)
-        return 0;
-
-    for (cursor_pos = mdata->cursor + 1; cursor_pos < l; cursor_pos++) {
-        struct string_formatted_chunk *chunk;
-
-        chunk = sol_vector_get(&mdata->chunks, cursor_pos);
-        if (chunk->type != STR_FORMAT_LITERAL) {
-            found = true;
-            break;
-        }
-    }
-
-    if (!found)
-        return 0;
-
-    mdata->cursor = cursor_pos;
-
-    mdata->state_changed = true;
-    mdata->blink_on = true;
-
-    string_formatted_force_imediate_format(mdata, true);
-    return string_formatted_format(node);
-}
-
-static int
-string_formatted_previous_set(struct sol_flow_node *node,
-    void *data,
-    uint16_t port,
-    uint16_t conn_id,
-    const struct sol_flow_packet *packet)
-{
-    struct string_formatted_data *mdata = data;
-    size_t cursor_pos = mdata->cursor;
-    bool found = false;
-
-    if (!mdata->enabled)
-        return 0;
-
-    while (cursor_pos) {
-        struct string_formatted_chunk *chunk;
-
-        cursor_pos--;
-        chunk = sol_vector_get(&mdata->chunks, cursor_pos);
-        if (chunk->type == STR_FORMAT_LITERAL)
-            continue;
-        else {
-            found = true;
-            break;
-        }
-    }
-
-    if (!found)
-        return 0;
-
-    mdata->cursor = cursor_pos;
-
-    mdata->state_changed = true;
-    mdata->blink_on = true;
-
-    string_formatted_force_imediate_format(mdata, true);
-    return string_formatted_format(node);
-}
-
-static int
-string_formatted_select_set(struct sol_flow_node *node,
-    void *data,
-    uint16_t port,
-    uint16_t conn_id,
-    const struct sol_flow_packet *packet)
-{
-    struct string_formatted_data *mdata = data;
-    int r;
-
-    if (!mdata->enabled)
-        return 0;
-
-    /* force new format with state changed and blink state on, so we
-     * always get the full output here */
-    string_formatted_force_imediate_format(mdata, false);
-    mdata->state_changed = true;
-    mdata->blink_on = true;
-
-    r = string_formatted_format(node);
-    SOL_INT_CHECK(r, < 0, r);
-
-    return sol_flow_send_string_slice_packet(node,
-        SOL_FLOW_NODE_TYPE_FORM_INT__OUT__SELECTED,
-        sol_buffer_get_slice(&mdata->formatted_value));
-}
-
-static int
-string_formatted_selected_set(struct sol_flow_node *node,
-    void *data,
-    uint16_t port,
-    uint16_t conn_id,
-    const struct sol_flow_packet *packet)
-{
-    struct string_formatted_data *mdata = data;
-    const char *value;
-    int r;
-
-    r = sol_flow_packet_get_string(packet, &value);
-    SOL_INT_CHECK(r, < 0, r);
-
-    r = string_formatted_selected_set_do(node, data, value);
-    SOL_INT_CHECK(r, < 0, r);
-
-    if (!mdata->enabled)
-        return 0;
-
-    string_formatted_force_imediate_format(mdata, true);
-    mdata->state_changed = true;
-    mdata->blink_on = true;
-
-    return string_formatted_format(node);
-}
-
-static int
-string_formatted_enabled_set(struct sol_flow_node *node,
-    void *data,
-    uint16_t port,
-    uint16_t conn_id,
-    const struct sol_flow_packet *packet)
-{
-    struct string_formatted_data *mdata = data;
-    bool value;
-    int r;
-
-    r = sol_flow_packet_get_boolean(packet, &value);
-    SOL_INT_CHECK(r, < 0, r);
-
-    mdata->enabled = value;
-
-    return 0;
 }
 
 #include "form-gen.c"
