+++ src/modules/flow/format/Makefile
@@ -1,0 +1,4 @@
+obj-$(FLOW_NODE_TYPE_FORMAT) += format.mod
+obj-format-$(FLOW_NODE_TYPE_FORMAT) := format.json format.o string-format.o ../form/form-common.o
+obj-format-$(FLOW_NODE_TYPE_FORMAT)-type := flow
+obj-format-$(FLOW_NODE_TYPE_FORMAT)-extra-ldflags := -lm
