+++ src/test-fbp/converter-string-float.fbp
@@ -17,7 +17,7 @@
 const_float(constant/float:value=3.1415)
 const_float_str(constant/string:value="3.141500")
 
-const_float OUT -> IN float_to_str(converter/float-to-string) OUT -> IN[0] str_cmp(string/compare)
+const_float OUT -> IN float_to_str(converter/float-to-string-format) OUT -> IN[0] str_cmp(string/compare)
 const_float_str OUT -> IN[1] str_cmp EQUAL -> RESULT float_converts_to_string(test/result)
 
 float_to_str OUT -> IN str_to_float(converter/string-to-float)
@@ -25,29 +25,54 @@ str_to_float OUT -> IN[0] is_float_equal(float/equal)
 const_float OUT -> IN[1] is_float_equal OUT -> RESULT string_converts_to_float(test/result)
 
 #stressing float formatting
-_(constant/float:value=3) OUT -> IN _(converter/float-to-string:format_spec="{:}") OUT -> IN[0] fmt_test_cmp_33(string/compare)
-_(constant/string:value="3.000000") OUT -> IN[1] fmt_test_cmp_33 EQUAL -> RESULT _(test/result)
+_(constant/float:value=3) OUT -> IN _(converter/float-to-string-format:format_spec="{:}") OUT -> IN[0] fmt_test_cmp_01(string/compare)
+_(constant/string:value="3.000000") OUT -> IN[1] fmt_test_cmp_01 EQUAL -> RESULT _(test/result)
 
-_(constant/float:value=DBL_MAX) OUT -> IN _(converter/float-to-string:format_spec="{:e}") OUT -> IN[0] fmt_test_cmp_34(string/compare)
-_(constant/string:value="1.797693e+308") OUT -> IN[1] fmt_test_cmp_34 EQUAL -> RESULT _(test/result)
+_(constant/float:value=DBL_MAX) OUT -> IN _(converter/float-to-string-format:format_spec="{:e}") OUT -> IN[0] fmt_test_cmp_02(string/compare)
+_(constant/string:value="1.797693e+308") OUT -> IN[1] fmt_test_cmp_02 EQUAL -> RESULT _(test/result)
 
-_(constant/float:value=DBL_MAX) OUT -> IN _(converter/float-to-string:format_spec="{:E}") OUT -> IN[0] fmt_test_cmp_35(string/compare)
-_(constant/string:value="1.797693E+308") OUT -> IN[1] fmt_test_cmp_35 EQUAL -> RESULT _(test/result)
+_(constant/float:value=DBL_MAX) OUT -> IN _(converter/float-to-string-format:format_spec="{:E}") OUT -> IN[0] fmt_test_cmp_03(string/compare)
+_(constant/string:value="1.797693E+308") OUT -> IN[1] fmt_test_cmp_03 EQUAL -> RESULT _(test/result)
 
-_(constant/float:value=3.1415) OUT -> IN _(converter/float-to-string:format_spec="{:g}") OUT -> IN[0] fmt_test_cmp_36(string/compare)
-_(constant/string:value="3.1415") OUT -> IN[1] fmt_test_cmp_36 EQUAL -> RESULT _(test/result)
+_(constant/float:value=3.1415) OUT -> IN _(converter/float-to-string-format:format_spec="{:g}") OUT -> IN[0] fmt_test_cmp_04(string/compare)
+_(constant/string:value="3.1415") OUT -> IN[1] fmt_test_cmp_04 EQUAL -> RESULT _(test/result)
 
-_(constant/float:value=30000000) OUT -> IN _(converter/float-to-string:format_spec="{:g}") OUT -> IN[0] fmt_test_cmp_37(string/compare)
-_(constant/string:value="3e+07") OUT -> IN[1] fmt_test_cmp_37 EQUAL -> RESULT _(test/result)
+_(constant/float:value=30000000) OUT -> IN _(converter/float-to-string-format:format_spec="{:g}") OUT -> IN[0] fmt_test_cmp_05(string/compare)
+_(constant/string:value="3e+07") OUT -> IN[1] fmt_test_cmp_05 EQUAL -> RESULT _(test/result)
 
-_(constant/float:value=30000000) OUT -> IN _(converter/float-to-string:format_spec="{:G}") OUT -> IN[0] fmt_test_cmp_38(string/compare)
-_(constant/string:value="3E+07") OUT -> IN[1] fmt_test_cmp_38 EQUAL -> RESULT _(test/result)
+_(constant/float:value=30000000) OUT -> IN _(converter/float-to-string-format:format_spec="{:G}") OUT -> IN[0] fmt_test_cmp_06(string/compare)
+_(constant/string:value="3E+07") OUT -> IN[1] fmt_test_cmp_06 EQUAL -> RESULT _(test/result)
 
-_(constant/float:value=30000000) OUT -> IN _(converter/float-to-string:format_spec="{:,f}") OUT -> IN[0] fmt_test_cmp_39(string/compare)
-_(constant/string:value="30,000,000.000000") OUT -> IN[1] fmt_test_cmp_39 EQUAL -> RESULT _(test/result)
+_(constant/float:value=30000000) OUT -> IN _(converter/float-to-string-format:format_spec="{:,f}") OUT -> IN[0] fmt_test_cmp_07(string/compare)
+_(constant/string:value="30,000,000.000000") OUT -> IN[1] fmt_test_cmp_07 EQUAL -> RESULT _(test/result)
 
-_(constant/float:value=3.1415) OUT -> IN _(converter/float-to-string:format_spec="{:%}") OUT -> IN[0] fmt_test_cmp_40(string/compare)
-_(constant/string:value="314.150000%") OUT -> IN[1] fmt_test_cmp_40 EQUAL -> RESULT _(test/result)
+_(constant/float:value=3.1415) OUT -> IN _(converter/float-to-string-format:format_spec="{:%}") OUT -> IN[0] fmt_test_cmp_08(string/compare)
+_(constant/string:value="314.150000%") OUT -> IN[1] fmt_test_cmp_08 EQUAL -> RESULT _(test/result)
 
-_(constant/float:value=3.1415) OUT -> IN _(converter/float-to-string:format_spec="{val:07n} ★ {min:.3f} ★ {max:G} ★ {step:%}") OUT -> IN[0] fmt_test_cmp_41(string/compare)
-_(constant/string:value="03.1415 ★ -179769313486231570814527423731704356798070567525844996598917476803157260780028538760589558632766878171540458953514382464234321326889464182768467546703537516986049910576551282076245490090389328944075868508455133942304583236903222948165808559332123348274797826204144723168738177180919299881250404026184124858368.000 ★ 1.79769E+308 ★ 0.000000%") OUT -> IN[1] fmt_test_cmp_41 EQUAL -> RESULT _(test/result)
+_(constant/float:value=3.1415) OUT -> IN _(converter/float-to-string-format:format_spec="{val:07n} ★ {min:.3f} ★ {max:G} ★ {step:%}") OUT -> IN[0] fmt_test_cmp_09(string/compare)
+_(constant/string:value="03.1415 ★ -179769313486231570814527423731704356798070567525844996598917476803157260780028538760589558632766878171540458953514382464234321326889464182768467546703537516986049910576551282076245490090389328944075868508455133942304583236903222948165808559332123348274797826204144723168738177180919299881250404026184124858368.000 ★ 1.79769E+308 ★ 0.000000%") OUT -> IN[1] fmt_test_cmp_09 EQUAL -> RESULT _(test/result)
+
+#float formatting for non-rich-format
+_(constant/float:value=3) OUT -> IN _(converter/float-to-string) OUT -> IN[0] fmt_test_cmp_10(string/compare)
+_(constant/string:value="3.000000") OUT -> IN[1] fmt_test_cmp_10 EQUAL -> RESULT _(test/result)
+
+_(constant/float:value=3.23000) OUT -> IN _(converter/float-to-string:format_flags="#",format_conversion_specifier="g") OUT -> IN[0] fmt_test_cmp_11(string/compare)
+_(constant/string:value="3.23000") OUT -> IN[1] fmt_test_cmp_11 EQUAL -> RESULT _(test/result)
+
+_(constant/float:value=3.2) OUT -> IN _(converter/float-to-string:format_flags="0",format_field_width="5",format_conversion_specifier="g") OUT -> IN[0] fmt_test_cmp_12(string/compare)
+_(constant/string:value="003.2") OUT -> IN[1] fmt_test_cmp_12 EQUAL -> RESULT _(test/result)
+
+_(constant/float:value=3) OUT -> IN _(converter/float-to-string:format_flags="-",format_field_width="13",format_conversion_specifier="f") OUT -> IN[0] fmt_test_cmp_13(string/compare)
+_(constant/string:value="3.000000     ") OUT -> IN[1] fmt_test_cmp_13 EQUAL -> RESULT _(test/result)
+
+_(constant/float:value=3) OUT -> IN _(converter/float-to-string:format_flags=" ",format_conversion_specifier="f") OUT -> IN[0] fmt_test_cmp_14(string/compare)
+_(constant/string:value=" 3.000000") OUT -> IN[1] fmt_test_cmp_14 EQUAL -> RESULT _(test/result)
+
+_(constant/float:value=3) OUT -> IN _(converter/float-to-string:format_flags="+",format_conversion_specifier="f") OUT -> IN[0] fmt_test_cmp_15(string/compare)
+_(constant/string:value="+3.000000") OUT -> IN[1] fmt_test_cmp_15 EQUAL -> RESULT _(test/result)
+
+_(constant/float:value=3) OUT -> IN _(converter/float-to-string:format_field_width="5",format_precision=".2",format_conversion_specifier="f") OUT -> IN[0] fmt_test_cmp_16(string/compare)
+_(constant/string:value=" 3.00") OUT -> IN[1] fmt_test_cmp_16 EQUAL -> RESULT _(test/result)
+
+# _(constant/float:value=3) OUT -> IN _(converter/float-to-string:format_prefix="prefix",format_suffix="suffix",format_field_width="5",format_precision=".2",format_conversion_specifier="f") OUT -> IN[0] fmt_test_cmp_17(string/compare)
+# _(constant/string:value="prefix 3.00suffix") OUT -> IN[1] fmt_test_cmp_17 EQUAL -> RESULT _(test/result)
